rp/pyoidc-rp/app/requirements.txt
Beaker==1.11.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
defusedxml==0.6.0
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-pyoidc==3.4.0
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
oic==1.1.2
pycparser==2.20
pycryptodomex==3.9.7
pyjwkest==1.4.2
python-dotenv==0.13.0
requests==2.23.0
six==1.15.0
typing-extensions==3.7.4.2
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
zipp==3.1.0
